[IMP] account_invoice_inter_company: move method _check_intercompany_product to account.move.line
