+++ account_invoice_inter_company/__init__.py
@@ -1,5 +1,1 @@
-# Copyright 2013-2014 Odoo SA
-# Copyright 2015-2017 Chafique Delli <chafique.delli@akretion.com>
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
 from . import models
