+++ account_invoice_inter_company/models/__init__.py
@@ -1,7 +1,4 @@
-# Copyright 2013-2014 Odoo SA
-# Copyright 2015-2017 Chafique Delli <chafique.delli@akretion.com>
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
 from . import account_move
+from . import account_move_line
 from . import res_config_settings
 from . import res_company
